Remove unused `rpc-client` library, clean it up a little
